medium link is largely unhelpful
This is a nice little example repo. It is a helpful piece of code for anyone learning rust that wants to see some working code and start hacking on it.

Some of the most talented people make the worst teachers, because they can't see how difficult it is to learn when everything seems to obvious to them.

The medium article by comparison is difficult to follow at best and incoherent gibberish at worst. If we compare it with https://pwy.io/en/posts/learning-to-fly-pt1/ it becomes obvious that the medium post is deficient as education for others.
I estimate that the target audience for that medium article is someone that has read The Rust Book and maybe has just finished rustlings. From the medium post they may be encouraged to abandon rust entirely.